Prompt rewrite, updated krb command, misc.
Prompt is now based on PS1 and control sequences, prompt_command
is only used to trigger the refresh of PS1.
Added 'krb -v' to get the name of the current principial, used in
the prompt.
Removed depend on tput.
Optimized git and svn check.
Print number of background jobs in right prompt.